// lib/content/types.ts
export interface CourseManagementInfo {
  id: string;
  title: string;
  sourceLanguage: string;
  targetLanguage: string;
  level: string;
  visibility: string | null;
  createdBy: string | null;
  units: {
    id: string;
    title: string;
    icon: string;
    visibility: string | null;
    lessonCount: number;
    questionType?: QuestionTypeInfo | null;
  }[];
}

export interface AvailableUnitForCourse {
  id: string;
  title: string;
  icon: string;
  targetLanguage: string;
  level: string | null;
  lessonCount: number;
  questionType?: QuestionTypeInfo | null;
}